third_party/WebKit/LayoutTests/inspector/report-API-errors-expected.txt

Tests that InspectorBackendStub is catching incorrect arguments.

Protocol Error: Invalid type of argument 'userAgent' for method 'Network.setUserAgentOverride' call. It must be 'string' but it is 'number'.
Protocol Error: Invalid number of arguments for method 'Network.setUserAgentOverride' call. It must have the following arguments '[{"name":"userAgent","type":"string","optional":false}]'.
Protocol Error: Optional callback argument for method 'Network.setUserAgentOverride' call must be a function but its type is 'string'.
Protocol Error: the message wrongDomain.something-strange is for non-existing domain 'wrongDomain': {"method":"wrongDomain.something-strange","params":{}}
Protocol Error: Attempted to dispatch an unspecified method 'Inspector.something-strange': {"method":"Inspector.something-strange","params":{}}
